refactor: move listing page to first architecture level
