+++ packages/manager/modules/veeam-cloud-connect/src/details/component.js
