+++ packages/manager/modules/veeam-cloud-connect/src/details/index.less
