+++ packages/manager/modules/veeam-cloud-connect/src/details/service.js
