+++ packages/manager/modules/veeam-cloud-connect/src/details/translations/Messages_de_DE.json
