+++ packages/manager/modules/veeam-cloud-connect/src/details/translations/Messages_es_ES.json
