+++ packages/manager/modules/veeam-cloud-connect/src/details/translations/Messages_fr_CA.json
