+++ packages/manager/modules/veeam-cloud-connect/src/details/translations/Messages_it_IT.json
