+++ packages/manager/modules/veeam-cloud-connect/src/details/translations/Messages_pt_PT.json
