--- packages/manager/modules/veeam-cloud-connect/src/template.html
@@ -1,3 +1,0 @@
-<div class="cui-page veeam">
-    <div class="veeam-message" data-ui-view="veeamContainer"></div>
-</div>
