+++ packages/manager/modules/veeam-cloud-connect/src/translations/Messages_fr_FR.json
@@ -1,45 +1,3 @@
 {
-  "veeam_tab_dashboard": "Accueil",
-  "veeam_tab_storage": "Espace de stockage",
-  "veeam_quota": "Quota",
-  "veeam_infos_general_load_error": "Erreur lors du chargement des informations générales.",
-  "veeam_infos_configuration_load_error": "Erreur lors du chargement des informations de configuration.",
-  "veeam_infos_subscription_load_error": "Erreur lors du chargement des informations d'abonnement.",
-  "veeam_offer_prices_load_error": "Erreur lors du chargement des offres et de leur prix.",
-  "veeam_orderable_offer_load_error": "Erreur lors du chargement des offres disponibles.",
-  "veeam_update_offer_post_success": "Le bon de commande <a href=\"{{orderUrl}}\" target=\"_blank\">{{orderId}}</a> a été créé avec succès.",
-  "veeam_update_offer_post_error": "Erreur lors de la création du bon de commande.",
-  "veeam_add_repository_pending": "La création de votre espace de stockage est en cours.",
-  "veeam_add_repository_error": "Erreur lors de la création de votre espace de stockage.",
-  "veeam_add_repository_success": "Votre espace de stockage a été créé.",
-  "veeam_infos_subscription_renew_manual_description": "Manuel",
-  "veeam_infos_subscription_renew_automatic_description": "Automatique",
-  "veeam_table_storage": "Espace de stockage",
-  "veeam_table_usage": "Utilisation",
-  "veeam_table_datacenter": "Datacenter de réplication",
-  "veeam_table_action": "Action",
-  "veeam_storage_list_load_error": "Erreur lors du chargement des espaces de stockage.",
-  "veeam_storage_update_quota_pending": "La mise à jour de votre quota de stockage est en cours.",
-  "veeam_storage_update_quota_error": "Erreur lors la mise à jour de votre quota de stockage.",
-  "veeam_storage_update_quota_success": "La mise à jour du quota de stockage est terminée.",
-  "veeam_tasks_success": "Les modifications ont été prise en compte.",
-  "veeam_tasks_pending": "Tâches en cours",
-  "veeam_action_upgrade_offer_no_upgrade": "L'offre {{offer}} ne peut être augmenter pour le moment.",
-  "veeam_action_add_storage_unavailable_offer": "L'offre {{offer}} ne vous permet pas d'ajouter des espaces de stockage.",
-  "veeam_action_add_storage_unavailable_storage_number": "Vous avez atteint le nombre maximal d'espace de stockage. L'ajout est désactivé.",
-  "veeam_action_add_storage_unavailable_usage_percentage": "Impossible d'ajouter un espace de stockage si vous n'utilisez pas au moins {{usagePercentage}}% de votre espace disponible.",
-  "veeam_common_action_unavailable": "Action indisponible",
-  "veeam_common_cancel": "Annuler",
-  "veeam_common_edit": "Modifier",
-  "veeam_common_manage": "Gérer",
-  "veeam_common_term_and_contracts": "Termes et contrats",
-  "veeam_common_term_and_contracts_accept": "Je certifie avoir lu et accepté les termes et contrats.",
-  "veeam_common_tile_subscription": "Abonnement",
-  "veeam_common_tile_subscription_contact": "Contacts",
-  "veeam_common_tile_subscription_contact_admin": "Administrateur",
-  "veeam_common_tile_subscription_contact_billing": "Facturation",
-  "veeam_common_tile_subscription_contact_technical": "Technique",
-  "veeam_common_tile_subscription_creation_date": "Date de création",
-  "veeam_common_tile_subscription_offer": "Offre",
-  "veeam_common_tile_subscription_autorenew": "Renouvellement automatique"
+  "veeam_cc_title": "Veeam Cloud Connect"
 }
