--- packages/manager/modules/veeam-cloud-connect/src/veeam/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "veeam_cc_title": "Veeam Cloud Connect"
-}
